requirements.txt
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.38
